requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
playwright
asyncio
aiohttp
python-dotenv
openai
tqdm
pydantic
reportlab
langchain
langgraph
faiss-cpu
langchain_openai
fpdf2